pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.0.1",
    "python-telegram-bot==13.15",
    "telegram>=0.0.1",
]
